+++ grails-app/domain/org/powertac/common/Tariff.groovy
@@ -42,10 +42,10 @@ import org.powertac.common.enumerations.PowerType
  * to call it inside the constructor for some reason.</p>
  * @author John Collins
  */
-class Tariff 
+class Tariff
 {
   // ----------- State enumeration --------------
-  
+
   enum State
   {
     PENDING, OFFERED, ACTIVE, WITHDRAWN, KILLED, INACTIVE
@@ -58,45 +58,55 @@ class Tariff
 
   /** The Tariff spec*/
   TariffSpecification tariffSpec
-  
+
   /** The broker behind this tariff */
   Broker broker
-  
+
   /** Last date new subscriptions will be accepted */
   Instant expiration
-  
+
   /** Current state of this Tariff */
   State state = State.PENDING
-  
+
   /** ID of Tariff that supersedes this Tariff */
   Tariff isSupersededBy
 
   /** Tracks the realized price for variable-rate tariffs. */
   Double totalCost = 0.0
   Double totalUsage = 0.0
-  
+
   /** Records the date when the Tariff was first offered */
   Instant offerDate
-  
+
   /** Maximum future interval over which price can be known */
   Duration maxHorizon // TODO lazy instantiation?
-  
+
   /** True if the maps are keyed by hour-in-week rather than hour-in-day */
   Boolean isWeekly = false
   Boolean analyzed = false
-  
-  // map is an array, indexed by tier-threshold and hour-in-day/week
-  //List< BigDecimal > tiers = []
-  //List< List< BigDecimal > > rateMap = []
 
-  static transients = ["realizedPrice", "usageCharge", "expired", "revoked", "timeService",
-                       "covered", "minDuration", "powerType", "signupPayment", 
-                       "earlyWithdrawPayment", "periodicPayment"]
-  
+  // map is an array, indexed by tier-threshold and hour-in-day/week
+  //List< double > tiers = []
+  //List< List< double > > rateMap = []
+
+  static transients = [
+    "realizedPrice",
+    "usageCharge",
+    "expired",
+    "revoked",
+    "timeService",
+    "covered",
+    "minDuration",
+    "powerType",
+    "signupPayment",
+    "earlyWithdrawPayment",
+    "periodicPayment"
+  ]
+
   static auditable = true
-  
+
   //static hasMany = [subscriptions:TariffSubscription]
-  
+
   static constraints = {
     specId(nullable: false, blank: false)
     broker(nullable:false)
@@ -104,8 +114,8 @@ class Tariff
     state(nullable: false)
     isSupersededBy(nullable: true)
     maxHorizon(nullable:true)
- }
-  
+  }
+
   //static mapping = {
   //  id (generator: 'assigned')
   //}
@@ -122,7 +132,7 @@ class Tariff
     offerDate = timeService.getCurrentTime()
     this.save()
     tariffSpec.getSupersedes().each {
-      Tariff supersededTariff = Tariff.findBySpecId(it) 
+      Tariff supersededTariff = Tariff.findBySpecId(it)
       supersededTariff.isSupersededBy = this
       supersededTariff.save()
     }
@@ -131,7 +141,7 @@ class Tariff
     broker.addToTariffs(this)
     broker.save()
   }
-  
+
   /**
    * Adds a new HourlyCharge to its Rate. Returns true just
    * in case the operation was successful.
@@ -150,39 +160,39 @@ class Tariff
     else
       return totalCost / totalUsage
   }
-  
+
   /** Pass-through for TariffSpecification.minDuration */
   long getMinDuration ()
   {
     tariffSpec.minDuration
   }
-  
+
   /** Type of power covered by this tariff */
   PowerType getPowerType ()
   {
     tariffSpec.powerType
   }
-  
+
   /** One-time payment for subscribing to tariff, positive for payment
    *  from customer, negative for payment to customer. */
-  BigDecimal getSignupPayment ()
+  double getSignupPayment ()
   {
     tariffSpec.signupPayment
   }
-  
+
   /** Payment from customer to broker for canceling subscription before
    *  minDuration has elapsed. */
-  BigDecimal getEarlyWithdrawPayment ()
+  double getEarlyWithdrawPayment ()
   {
     tariffSpec.earlyWithdrawPayment
   }
-  
+
   /** Flat payment per period for two-part tariffs */
-  BigDecimal getPeriodicPayment ()
+  double getPeriodicPayment ()
   {
     tariffSpec.periodicPayment
   }
-  
+
   /**
    * Adds periodic payments to the total cost, so realized price includes it.
    */
@@ -211,7 +221,7 @@ class Tariff
     }
     return amt
   }
-  
+
   /** 
    * Returns the usage charge for a single customer using an amount of 
    * energy at some time in 
@@ -225,7 +235,7 @@ class Tariff
   {
     // start by retrieving the rate map
     //def rateMap = tariffRateService.getRateMap(this)
-    
+
     // first, get the time index
     DateTime dt = new DateTime(when, DateTimeZone.UTC)
     int di = dt.getHourOfDay()
@@ -282,7 +292,7 @@ class Tariff
       return result
     }
   }
-  
+
   /**
    * True just in case the current time is past the expiration date
    * of this Tariff.
@@ -296,7 +306,7 @@ class Tariff
       return timeService.getCurrentTime().millis >= expiration.millis
     }
   }
-  
+
   /**
    * True just in case the set of Rates cover all the possible hour
    * and tier slots. If false, then there is some combination of hour
@@ -308,7 +318,7 @@ class Tariff
     for (tier in 0..<tiers.size()) {
       for (hour in 0..<(isWeekly? 24 * 7: 24)) {
         //def cell = rateMap[tier][hour]
-        //println "cell: ${cell}" 
+        //println "cell: ${cell}"
         //if (cell == null) {
         //  return false
         //}
@@ -319,7 +329,7 @@ class Tariff
     }
     return true
   }
-  
+
   /**
    * True just in case this tariff has been revoked.
    */
@@ -360,14 +370,14 @@ class Tariff
     tariffRateService.sortTiers(this)
     def tiers = tariffRateService.getTiers(this)
     log.info "tariff ${this.id}, tiers: ${tiers}"
-    
+
     // Next, fill in the tierIndexMap, which maps tier thresholds to
     // array indices. Remember that there's always a 0.0 tier.
     int tidx = 0
     tiers.each { threshold ->
       tierIndexMap[(threshold)] = tidx++
     }
-    
+
     // Now we can compute the sort keys. Note that the lowest-priority
     // rates will sort first.
     def annotatedRates = [:] as TreeMap
